Revert "Attestation changes"
