+++ app/allocation/[category]/attestation/AttestationSuccessModal.tsx
@@ -1,55 +1,40 @@
 import Image from 'next/image';
-import React, { useState } from 'react';
-import { UpdateBallotButton } from '../components/UpdateBallotButton';
+import React from 'react';
+import { ExternalLinkIcon } from '@/public/assets/icon-components/ExternalLink';
+
 interface Props {
   link: string
   onClose: () => void
 }
 
-const AttestationSuccessModal: React.FC<Props> = ({ link, onClose }) => {
-  const [hideAttestation, setHideAttestation] = useState(false);
+const AttestationSuccessModal: React.FC<Props> = ({ link }) => {
   return (
-    <div className={`${hideAttestation ? 'hidden' : 'mx-auto flex max-w-md flex-col items-center gap-6 overflow-hidden rounded-lg bg-white bg-ballot bg-no-repeat p-6 py-10 text-center shadow-lg'}`}>
+    <div className="mx-auto flex max-w-md flex-col items-center gap-6 overflow-hidden rounded-lg bg-white bg-ballot bg-no-repeat p-6 py-12 text-center shadow-lg">
       <Image
-        src="/assets/images/op-voting-char.svg"
+        src="/assets/images/sunny.svg"
         alt="Celebration"
-        width={300}
-        height={150}
+        width={160}
+        height={125}
         className="mx-auto mb-6"
       />
-
-      <div className="items-around flex flex-col gap-4">
-        <h2 className="whitespace-nowrap text-2xl font-bold text-dark-500">
-          Attestation done successfully!!
-        </h2>
-        <p className="text-gray-400">
-          It&39;s time to celebrate!! Your votes have been attested successfully.
-        </p>
-        <p className="text-gray-400">
-          You can update your ballot on Optimism.
-        </p>
+      <h2 className="whitespace-nowrap text-2xl font-bold text-dark-500">
+        Attestation done successfully!!
+      </h2>
+      <p className="text-slate-700">
+        It's time to celebrate!! Your have votes have been attested successfully.
+      </p>
+      <div className="flex w-3/5 flex-col gap-1">
+        <a href={link} className="w-full" target="_blank">
+          <button
+            className="mb-2 flex w-full items-center justify-center gap-2 rounded-lg px-4 py-3
+          text-primary"
+          >
+            View Transaction
+            <ExternalLinkIcon />
+          </button>
+        </a>
       </div>
-      <div>
-        <div onClick={() => setHideAttestation(true)}>
-          <UpdateBallotButton closeAttestationModal={() => onClose()} />
-        </div>
-      </div>
-      <a href={link} className="w-full" target="_blank">
-        <button
-          className="text-primary underline"
-        >
-          View Transaction
-        </button>
-      </a>
-
-      <button
-        className="text-gray-600"
-        onClick={onClose}
-      >
-        Close
-      </button>
     </div>
-
   );
 };
